+++ src/helper/chunkArray.js
@@ -1,10 +1,11 @@
+var is = require('../helper/is');
 // chunk an array (arr) into arrays of (size) items, returns an array of arrays, or an empty array on invalid input
 module.exports = function (arr, size) {
   var returnVal = [],
     x = 0,
     len = arr.length;
 
-  if (size < 1 || !Twig.lib.is("Array", arr)) {
+  if (size < 1 || !is("Array", arr)) {
     return [];
   }