+++ changes.md
@@ -2,3 +2,7 @@
 - `Twig.indexOf` -> `require('./helper/indexOf')`
 - `Twig.forEach` -> `require('./helper/forEach')`
 - `Twig.merge` -> `require('./helper/merge')`
+- `Twig.lib.is` -> `require('./helper/is')`
+- `Twig.lib.copy` -> `require('./helper/copy')`
+- `Twig.lib.extend` -> `require('./helper/extend')`
+- `Twig.lib.replaceAll` -> `require('./helper/replaceAll')`